Move weather settings to own dir
